feat: load initial characters into vocabulary

+++ include/args.h
@@ -1,11 +1,12 @@
 #ifndef ARGS_H_
     #define ARGS_H_
     #include <limits.h>
+    #include <stdint.h>
     #define DEFAULT_VOCABULARY_CAPACITY ULLONG_MAX
 
 #define OPTIONAL_ARGS \
     OPTIONAL_STRING_ARG(input, "", "-i", "input", "The input to encode.") \
-    OPTIONAL_ULONG_LONG_ARG(capacity, 1000, "-c", "capacity", "The vocabulary capacity")
+    OPTIONAL_ULONG_LONG_ARG(capacity, (uint64_t)1000, "-c", "capacity", "The vocabulary capacity")
 
 #define BOOLEAN_ARGS \
     BOOLEAN_ARG(use_stdin, "-s", "Use stdin as input.") \
